requirements.txt
telethon
slack_sdk
python-dotenv
PyQt6
PyQt6-tools
PyQt6-sip
requests
